Move application submission to POST to /cas2/submissions
Note that the `applicationId` is now sent as the `applicationId`
url param rather than being in the path.
